Split reducers.
